BAU: move type modules to standard imports for integration tests
Integration tests have been long blocked on the mix of JavaScript/
TypeScript and a quirk with `ts-node`. Move imports from a compile time
module into direct references to allow the server to run with both
`ts-node` and compiled by `tsc.

This is all a stop gap before the build process is redesigned once all
files have been moved over to TypeScript.

Add supertest to allow simple assertions against the running server.
