+++ package.json
@@ -79,8 +79,11 @@
     "@types/json2csv": "4.5.0",
     "@types/lodash": "4.14.136",
     "@types/mocha": "7.0.2",
+    "@types/morgan": "1.7.37",
+    "@types/multer": "1.4.2",
     "@types/node": "13.13.5",
     "@types/passport": "1.0.0",
+    "@types/qs": "6.9.2",
     "@types/react": "16.9.35",
     "@types/react-dom": "16.9.8",
     "@types/stripe": "6.25.8",
@@ -89,9 +92,6 @@
     "@types/webpack-node-externals": "1.7.1",
     "@typescript-eslint/eslint-plugin": "2.32.0",
     "@typescript-eslint/parser": "2.32.0",
-    "@types/morgan": "1.7.37",
-    "@types/multer": "1.4.2",
-    "@types/qs": "6.9.2",
     "chai": "4.2.0",
     "chai-spies": "1.0.0",
     "copyfiles": "2.2.0",
@@ -102,14 +102,15 @@
     "mocha": "7.1.2",
     "mockery": "2.1.0",
     "node-sass": "4.14.1",
-    "sinon": "9.0.2",
     "react": "16.13.1",
     "react-dom": "16.13.1",
     "react-inlinesvg": "1.2.0",
     "react-spring": "8.0.27",
+    "sinon": "9.0.2",
     "source-map-loader": "0.2.4",
+    "supertest": "^4.0.2",
     "ts-loader": "7.0.4",
-    "ts-node": "8.10.1",
+    "ts-node": "^8.10.2",
     "tsc-watch": "4.2.3",
     "typescript": "3.8.3",
     "url-loader": "4.1.0",
