--- src/web/modules/transactions/@types/reports.d.ts
@@ -1,19 +1,0 @@
-declare module 'reports' {
-
-  export interface PaymentsByStatusReport {
-    success: number;
-    declined: number;
-    timedout: number;
-    cancelled: number;
-    error: number;
-    created: number;
-    started: number;
-    submitted: number;
-    capturable: number;
-  }
-
-  export interface PaymentsReport {
-    count: number;
-    gross_amount: number;
-  }
-}
